Don't remove resources from the DB during HA restart
This was broken by 3b91d100a6c6424961e1ec06cf7afd9a1c9cf49c. Resources
should be removed from the DB only when the stack is deleted.

+++ heat/engine/resources.py
@@ -190,19 +190,26 @@ class Resource(object):
             logger.exception('Delete %s', str(self))
             self.state_set(self.DELETE_FAILED, str(ex))
             return str(ex)
-        else:
-            try:
-                db_api.resource_get(self.stack.context, self.id).delete()
-            except exception.NotFound:
-                # Don't fail on delete if the db entry has
-                # not been created yet.
-                pass
-            except Exception as ex:
-                self.state_set(self.DELETE_FAILED)
-                logger.exception('Delete %s from DB' % str(self))
-                return str(ex)
 
-            self.state_set(self.DELETE_COMPLETE)
+        self.state_set(self.DELETE_COMPLETE)
+
+    def destroy(self):
+        '''
+        Delete the resource and remove it from the database.
+        '''
+        result = self.delete()
+        if result:
+            return result
+
+        try:
+            db_api.resource_get(self.stack.context, self.id).delete()
+        except exception.NotFound:
+            # Don't fail on delete if the db entry has
+            # not been created yet.
+            pass
+        except Exception as ex:
+            logger.exception('Delete %s from DB' % str(self))
+            return str(ex)
 
     def instance_id_set(self, inst):
         self.instance_id = inst
